Separate index.jsx into two modules to allow for testing
